Move git operations into separate module.
